windows dylib working

+++ src/bdjopt.h
@@ -102,7 +102,7 @@ typedef enum {
 void    bdjoptInit (void);
 void    bdjoptFree (void);
 void    *bdjoptGetData (ssize_t idx);
-long    bdjoptGetNum (ssize_t idx);
+ssize_t bdjoptGetNum (ssize_t idx);
 void    bdjoptSetNum (ssize_t idx, ssize_t value);
 
 #endif /* INC_BDJOPT_H */
